Save considerable image size by zerofilling
When deleting files on an ext4 system, they are of course not actually
deleted. Their blocks are only marked for reuse. That meant we were
keeping all of the deleted blocks from the system-image files. This was
about 400MB of xz-compressed data which wouldn't compress again and
greatly slowed down compressors. To make things worse, bmap-tools didn't
know the data was unused and caused it to be written out to the final
device. Ugh!

The zerofree tool fixes this by finding the blocks and zeroing them. A
further 'fallocate --dig-holes' then takes these zeroes and converts
them back into unused space on the host.

Another option was dd'ing a bunch of zeroes then deleting, but that
would cause the file to balloon to its full size first and slow down the
build. Finally, 'fstrim' could be used for this if the virtual disks
used by debos supported discard and used it correctly. They do not.